feat: add filter offers by creator_id and min_price

+++ app_offers/models.py
@@ -1,7 +1,7 @@
 from django.db import models
-from app_accounts.models import User 
+from app_accounts.models import User
 
- 
+
 class Offer(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     title = models.CharField(max_length=100)
@@ -9,11 +9,9 @@ class Offer(models.Model):
     description = models.TextField()
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
-    details = models.JSONField(null= True, blank=True)
-    min_price = models.IntegerField()
+    details = models.JSONField(null=True, blank=True)
+    min_price = models.FloatField()
     min_delivery_time = models.IntegerField()
-     
 
     def __str__(self):
-         return self.title
-
+        return self.title
